FakeR

Documentation for FakeR.

FakeR.c

FakeR.dim

FakeR.dir_create

FakeR.file_info

FakeR.grep

FakeR.grepl

FakeR.list_files

FakeR.range

FakeR.toupper

FakeR.@capture_output

FakeR.c — Method

R: c(1,2,3) vector constructor
julia: [1,2,3]

FakeR.dim — Method

R: dim() dimentions of array
julia: size

FakeR.dir_create — Method

R: dir.create(path, recursive = FALSE , mode = "0777").
create dictionary
dir_create
julia: mkpath, mkdir
Using mkpath so effectively recursive = TRUE always.
Also dir.create() inly takes single path, dir_create can take a vector

FakeR.file_info — Method

R: file.info(file)
returns data.frame of info on file
file_info
julia: stat
stat does not return dataframe, but this does

FakeR.grep — Method

R: grep(pattern, x, ignore.case = FALSE, value = FALSE, invert = FALSE)
return indices or values of match
julia: filter and occursin/contains

FakeR.grepl — Method

R: grepl(pattern, vector, ignore.case=false)
look for pattern in vector
julia: occursin/contains. Use string for the pattern
FakeR.grepl("^a", ["ale", "bale"])

FakeR.list_files — Function

R: list.files(path=".", pattern=NULL, all.files=FALSE, full.names=FALSE, recursive=FALSE, ignore.case=FALSE, inclue.dirs=FALSE, no..=FALSE)
list files
list_files. TODO: more keyword arguments
julia: readdir(path=".",; join=false, sort=true)

FakeR.range — Method

R: range(c(3,2,1)) == c(1,3)
Retuns a vector of the extreme values of a vector.
julia: extrema([3,2,1])
Nore range() in Julia has meaning similar to seq() in R

FakeR.toupper — Method

R: toupper(), tolower()
Uppercase or lowercase string or vector of strings
julia: uppercase, lowercase

FakeR.@capture_output — Macro

R: capture.output()
Capture the output of a printing command.
@capture_output print(a)
Returns the output of the expression as a string